+++ src/main.cpp
@@ -33,8 +33,6 @@ void IconPreview::updateShipGlow(SimplePlayer* glowPlayer) {
 
     glowPlayer->updatePlayerFrame(manager->getPlayerShip(), IconType::Ship);
     MoreIcons::updateSimplePlayer(glowPlayer, IconType::Ship);
-    glowPlayer->setColor(manager->colorForIdx(manager->getPlayerColor()));
-    glowPlayer->setSecondColor(manager->colorForIdx(manager->getPlayerColor2()));
     glowPlayer->setGlowOutline(manager->colorForIdx(manager->getPlayerGlowColor()));
     glowPlayer->enableCustomGlowColor(manager->colorForIdx(manager->getPlayerGlowColor()));
     if (!manager->getPlayerGlow()) {
@@ -46,8 +44,6 @@ void IconPreview::updateBirdGlow(SimplePlayer* glowPlayer) {
 
     glowPlayer->updatePlayerFrame(manager->getPlayerBird(), IconType::Ufo);
     MoreIcons::updateSimplePlayer(glowPlayer, IconType::Ufo);
-    glowPlayer->setColor(manager->colorForIdx(manager->getPlayerColor()));
-    glowPlayer->setSecondColor(manager->colorForIdx(manager->getPlayerColor2()));
     glowPlayer->setGlowOutline(manager->colorForIdx(manager->getPlayerGlowColor()));
     glowPlayer->enableCustomGlowColor(manager->colorForIdx(manager->getPlayerGlowColor()));
     if (!manager->getPlayerGlow()) {
@@ -59,8 +55,6 @@ void IconPreview::updateJetpackGlow(SimplePlayer* glowPlayer) {
 
     glowPlayer->updatePlayerFrame(manager->getPlayerJetpack(), IconType::Jetpack);
     MoreIcons::updateSimplePlayer(glowPlayer, IconType::Jetpack);
-    glowPlayer->setColor(manager->colorForIdx(manager->getPlayerColor()));
-    glowPlayer->setSecondColor(manager->colorForIdx(manager->getPlayerColor2()));
     glowPlayer->setGlowOutline(manager->colorForIdx(manager->getPlayerGlowColor()));
     glowPlayer->enableCustomGlowColor(manager->colorForIdx(manager->getPlayerGlowColor()));
     if (!manager->getPlayerGlow()) {
@@ -71,23 +65,95 @@ void IconPreview::updateJetpackGlow(SimplePlayer* glowPlayer) {
 // ------------------------------------
 // FUNCTIONS FOR SEPARATE DUAL ICONS FUNCTIONALITY
 // ------------------------------------
+void IconPreview::updateDualPreviewCube(SimplePlayer* player) {
+    auto manager = GameManager::sharedState();
+    auto dualMod = Loader::get()->getLoadedMod("weebify.separate_dual_icons");
 
+    int dualColor1 = dualMod->getSavedValue<int64_t>("color1");
+    int dualColor2 = dualMod->getSavedValue<int64_t>("color2");
+    bool dualHasGlow = dualMod->getSavedValue<bool>("glow");
+    int glowColor = dualMod->getSavedValue<int64_t>("colorglow");
 
+    int dualCubeFrame = dualMod->getSavedValue<int64_t>("cube");
 
+    player->updatePlayerFrame(dualCubeFrame, IconType::Cube);
+    MoreIcons::updateSimplePlayer(player, IconType::Cube, true);
+    player->setColor(manager->colorForIdx(dualColor1));
+    player->setSecondColor(manager->colorForIdx(dualColor2));
+    player->setGlowOutline(manager->colorForIdx(glowColor));
+    player->enableCustomGlowColor(manager->colorForIdx(glowColor));
+    if (!dualHasGlow) {
+        player->disableGlowOutline();
+    }
+}
+void IconPreview::updateDualShipGlow(SimplePlayer* glowPlayer) {
+    auto manager = GameManager::sharedState();
+    auto dualMod = Loader::get()->getLoadedMod("weebify.separate_dual_icons");
+
+    bool dualHasGlow = dualMod->getSavedValue<bool>("glow");
+    int glowColor = dualMod->getSavedValue<int64_t>("colorglow");
+    int dualShipFrame = dualMod->getSavedValue<int64_t>("ship");
+
+    glowPlayer->updatePlayerFrame(dualShipFrame, IconType::Ship);
+    MoreIcons::updateSimplePlayer(glowPlayer, IconType::Ship, true);
+    glowPlayer->setGlowOutline(manager->colorForIdx(glowColor));
+    glowPlayer->enableCustomGlowColor(manager->colorForIdx(glowColor));
+    if (!dualHasGlow) {
+        glowPlayer->disableGlowOutline();
+    }
+}
+void IconPreview::updateDualBirdGlow(SimplePlayer* glowPlayer) {
+    auto manager = GameManager::sharedState();
+    auto dualMod = Loader::get()->getLoadedMod("weebify.separate_dual_icons");
+
+    bool dualHasGlow = dualMod->getSavedValue<bool>("glow");
+    int glowColor = dualMod->getSavedValue<int64_t>("colorglow");
+    int dualBirdFrame = dualMod->getSavedValue<int64_t>("bird");
+
+    glowPlayer->updatePlayerFrame(dualBirdFrame, IconType::Ufo);
+    MoreIcons::updateSimplePlayer(glowPlayer, IconType::Ufo, true);
+    glowPlayer->setGlowOutline(manager->colorForIdx(glowColor));
+    glowPlayer->enableCustomGlowColor(manager->colorForIdx(glowColor));
+    if (!dualHasGlow) {
+        glowPlayer->disableGlowOutline();
+    }
+}
+void IconPreview::updateDualJetpackGlow(SimplePlayer* glowPlayer) {
+    auto manager = GameManager::sharedState();
+    auto dualMod = Loader::get()->getLoadedMod("weebify.separate_dual_icons");
+
+    bool dualHasGlow = dualMod->getSavedValue<bool>("glow");
+    int glowColor = dualMod->getSavedValue<int64_t>("colorglow");
+    int dualJetpackFrame = dualMod->getSavedValue<int64_t>("jetpack");
+
+    glowPlayer->updatePlayerFrame(dualJetpackFrame, IconType::Jetpack);
+    MoreIcons::updateSimplePlayer(glowPlayer, IconType::Jetpack, true);
+    glowPlayer->setGlowOutline(manager->colorForIdx(glowColor));
+    glowPlayer->enableCustomGlowColor(manager->colorForIdx(glowColor));
+    if (!dualHasGlow) {
+        glowPlayer->disableGlowOutline();
+    }
+}
 
 auto doGlowFix = Mod::get()->getSettingValue<bool>("glow-fix");
+auto enableReloadBtn = Mod::get()->getSettingValue<bool>("reload-btn");
 
 $on_mod(Loaded) {
     listenForSettingChanges("glow-fix", [](bool value) {
         doGlowFix = value;
+    });
+    listenForSettingChanges("reload-btn", [](bool value) {
+        enableReloadBtn = value;
     });
 }
 
 class $modify(IPGarageLayer, GJGarageLayer){
     struct Fields {
         SimplePlayer* m_previewPlayer = SimplePlayer::create(0);
-        SimplePlayer* m_dualPreview = SimplePlayer::create(0); //xdd
         SimplePlayer* fakeGlowDisplay = SimplePlayer::create(0);
+        // separate dual icons
+        SimplePlayer* m_dualPreview = SimplePlayer::create(0);
+        SimplePlayer* m_dualGlowDisplay = SimplePlayer::create(0);
     };
 
     static void onModify(auto& self) {
@@ -108,7 +174,17 @@ class $modify(IPGarageLayer, GJGarageLayer){
         float jPosX = m_playerObject->getPositionX() + 9.f;
         float jPosY = m_playerObject->getPositionY() + 6.f;
 
-        float jetpackCubeScale = 0.9f;
+        // dual icons stuff
+        bool isSeparateLoaded = Loader::get()->isModLoaded("weebify.separate_dual_icons");
+        auto SDL = Loader::get()->getLoadedMod("weebify.separate_dual_icons");
+        int lastDualMode = 0;
+        float dualPosX = 0.f;
+        float dualPosY = 0.f;
+        float dualShipPosY = 0.f;
+        float dualUfoPosY = 0.f;
+        float dualJetpackPosX = 0.f;
+        float dualJetpackPosY = 0.f;
+        bool dualHasGlow = false;
 
         if (lastSelectedIcon == IconType::Ship || lastSelectedIcon == IconType::Ufo || lastSelectedIcon == IconType::Jetpack) {
             fields->m_previewPlayer->setVisible(true);
@@ -126,15 +202,13 @@ class $modify(IPGarageLayer, GJGarageLayer){
 
         } else if (lastSelectedIcon == IconType::Jetpack) {
             fields->m_previewPlayer->setPosition({jPosX, jPosY});
-            fields->m_previewPlayer->setScale(jetpackCubeScale);
+            fields->m_previewPlayer->setScale(0.9f);
             
         } else {
             fields->m_previewPlayer->setPosition({posX, posY + 20.f});
         }
 
-        SimplePlayer* playerToSend = fields->m_previewPlayer;
-
-        IconPreview::updatePreviewCube(playerToSend);
+        IconPreview::updatePreviewCube(fields->m_previewPlayer);
 
         if (doGlowFix) {
             
@@ -176,16 +250,97 @@ class $modify(IPGarageLayer, GJGarageLayer){
         fields->m_previewPlayer->setID("preview-simpleplayer"_spr);
         this->addChild(fields->m_previewPlayer, -1);
 
+        if (isSeparateLoaded) {
+            lastDualMode = SDL->getSavedValue<int64_t>("lastmode");
+
+            SimplePlayer* p2Icon = as<SimplePlayer*>(this->getChildByID("player2-icon"));
+            dualPosX = p2Icon->getPositionX();
+            dualPosY = p2Icon->getPositionY();
+            dualShipPosY = p2Icon->getPositionY() + 16.f;
+            dualUfoPosY = p2Icon->getPositionY() + 8.5f;
+            dualJetpackPosX = p2Icon->getPositionX() + 9.f;
+            dualJetpackPosY = p2Icon->getPositionY() + 6.f;
+            dualHasGlow = SDL->getSavedValue<bool>("glow");
+
+            if (lastDualMode == 1 || lastDualMode == 3 || lastDualMode == 8) {
+                fields->m_dualPreview->setVisible(true);
+            } else {
+                fields->m_dualPreview->setVisible(false);
+            }
+
+            if (lastDualMode == 1) { // SHIP
+                fields->m_dualPreview->setPosition({dualPosX, dualShipPosY});
+                fields->m_dualPreview->setScale(0.88f);
+                p2Icon->setScale(1.6f);
+            } else if (lastDualMode == 3) { // UFO
+                fields->m_dualPreview->setPosition({dualPosX, dualUfoPosY});
+                fields->m_dualPreview->setScale(0.88f);
+                p2Icon->setScale(1.6f);
+            } else if (lastDualMode == 8) { // JETPACK
+                fields->m_dualPreview->setPosition({dualJetpackPosX, dualJetpackPosY});
+                fields->m_dualPreview->setScale(0.9f);
+                p2Icon->setScale(1.5f);
+            } else {
+                fields->m_dualPreview->setPosition({dualPosX, dualPosY + 20.f});
+            }
+
+            IconPreview::updateDualPreviewCube(fields->m_dualPreview);
+
+            if (doGlowFix) {
+                fields->m_dualGlowDisplay->setPosition({dualPosX, dualPosY});
+                fields->m_dualGlowDisplay->setID("fake-dual-glow-display"_spr);
+
+                if (lastDualMode == 1) {
+                    fields->m_dualGlowDisplay->setScale(1.6f);
+                    IconPreview::updateDualShipGlow(fields->m_dualGlowDisplay);
+
+                } else if (lastDualMode == 3) {
+                    fields->m_dualGlowDisplay->setScale(1.6f);
+                    IconPreview::updateDualBirdGlow(fields->m_dualGlowDisplay);
+
+                } else if (lastDualMode == 8) {
+                    fields->m_dualGlowDisplay->setScale(1.5f);
+                    IconPreview::updateDualJetpackGlow(fields->m_dualGlowDisplay);
+                }
+
+                fields->m_dualGlowDisplay->m_firstLayer->setOpacity(0);
+                fields->m_dualGlowDisplay->m_secondLayer->setVisible(false);
+                fields->m_dualGlowDisplay->m_birdDome->setVisible(false);
+                fields->m_dualGlowDisplay->m_detailSprite->setVisible(false);
+
+                if (lastDualMode == 1 || lastDualMode == 3 || lastDualMode == 8) {
+                    if (dualHasGlow) {
+                        fields->m_dualGlowDisplay->setVisible(true);
+                        fields->m_dualGlowDisplay->m_outlineSprite->setVisible(true);
+                        p2Icon->m_outlineSprite->setVisible(false);
+                    } else {
+                        fields->m_dualGlowDisplay->setVisible(false);
+                    }
+                }
+
+                this->addChild(fields->m_dualGlowDisplay, -2);
+            }
+
+            fields->m_dualPreview->setID("dual-preview-simpleplayer"_spr);
+            this->addChild(fields->m_dualPreview, -1);
+
+        }
+
         return true;
     }
 
     void updatePlayerColors() {
         GJGarageLayer::updatePlayerColors();
-        // THANK FUCKING GOD FOR THIS FUNCTION I DON'T HAVE TO DO A BUNCH OF SHIT TO UPDATE ON PALETTE-MENU-THINGY-WHATEVER CLOSING :yas:
 
         auto fields = m_fields.self();
         auto manager = GameManager::sharedState();
         auto lastChosenIcon = manager->m_playerIconType;
+
+        // dual icons stuff
+        bool isSeparateLoaded = Loader::get()->isModLoaded("weebify.separate_dual_icons");
+        auto SDL = Loader::get()->getLoadedMod("weebify.separate_dual_icons");
+        int lastDualMode = 0;
+        bool dualHasGlow = false;
 
         IconPreview::updatePreviewCube(fields->m_previewPlayer);
 
@@ -227,6 +382,61 @@ class $modify(IPGarageLayer, GJGarageLayer){
             }
 
         }
+        
+        if (isSeparateLoaded) {
+            lastDualMode = SDL->getSavedValue<int64_t>("lastmode");
+
+            dualHasGlow = SDL->getSavedValue<bool>("glow");
+            SimplePlayer* p2Icon = as<SimplePlayer*>(this->getChildByID("player2-icon"));
+
+            IconPreview::updateDualPreviewCube(fields->m_dualPreview);
+
+            if (doGlowFix) {
+                if (lastDualMode == 1) {
+                    fields->m_dualGlowDisplay->setScale(1.6f);
+                    IconPreview::updateDualShipGlow(fields->m_dualGlowDisplay);
+
+                } else if (lastDualMode == 3) {
+                    fields->m_dualGlowDisplay->setScale(1.6f);
+                    IconPreview::updateDualBirdGlow(fields->m_dualGlowDisplay);
+
+                } else if (lastDualMode == 8) {
+                    fields->m_dualGlowDisplay->setScale(1.5f);
+                    IconPreview::updateDualJetpackGlow(fields->m_dualGlowDisplay);
+                }
+
+                fields->m_dualGlowDisplay->m_firstLayer->setOpacity(0);
+                fields->m_dualGlowDisplay->m_secondLayer->setVisible(false);
+                fields->m_dualGlowDisplay->m_birdDome->setVisible(false);
+                fields->m_dualGlowDisplay->m_detailSprite->setVisible(false);
+
+                if (lastDualMode == 1 || lastDualMode == 3 || lastDualMode == 8) {
+                    if (dualHasGlow) {
+                        fields->m_dualGlowDisplay->setVisible(true);
+                        fields->m_dualGlowDisplay->m_outlineSprite->setVisible(true);
+                        if (p2Icon){
+                            p2Icon->m_outlineSprite->setVisible(false);
+                        } else {
+                            geode::log::debug("p2 icon not found! on glow enabled, something went wrong");
+                        }
+                    }
+                } else {
+                    fields->m_dualGlowDisplay->setVisible(false);
+                    if (p2Icon) {
+                        p2Icon->m_outlineSprite->setVisible(dualHasGlow);
+                    } else {
+                        geode::log::debug("p2 icon not found! on glow disabled, something went wrong");
+                    }
+                }
+
+                if (lastDualMode == 8) {
+                    fields->m_dualGlowDisplay->setScale(1.5f);
+                } else {
+                    fields->m_dualGlowDisplay->setScale(1.6f);
+                }
+            }
+
+        }
 
     }
 
@@ -247,6 +457,19 @@ class $modify(IPGarageLayer, GJGarageLayer){
         float jetpackPosX = m_playerObject->getPositionX() + 9.f;
         float jetpackPosY = m_playerObject->getPositionY() + 6.f;
         float jetpackScale = 0.9f;
+
+        // dual icons stuff
+        bool isSeparateLoaded = Loader::get()->isModLoaded("weebify.separate_dual_icons");
+        auto SDL = Loader::get()->getLoadedMod("weebify.separate_dual_icons");
+        int lastDualMode = 0;
+        float dualPosX = 0.f;
+        float dualPosY = 0.f;
+        float dualShipPosY = 0.f;
+        float dualUfoPosY = 0.f;
+        float dualJetpackPosX = 0.f;
+        float dualJetpackPosY = 0.f;
+        bool dualHasGlow = false;
+        SimplePlayer* p2Icon = nullptr;
 
         // this runs on selecting any mode just to make sure it updates properly
         IconPreview::updatePreviewCube(fields->m_previewPlayer);
@@ -327,6 +550,84 @@ class $modify(IPGarageLayer, GJGarageLayer){
             fields->m_previewPlayer->runAction(moveToJetpackPos);
             fields->m_previewPlayer->runAction(scaleToJetpackSize);
         }
+
+
+        if (isSeparateLoaded) {
+            lastDualMode = SDL->getSavedValue<int64_t>("lastmode");
+
+            p2Icon = static_cast<SimplePlayer*>(this->getChildByID("player2-icon"));
+            dualPosX = p2Icon->getPositionX();
+            dualPosY = p2Icon->getPositionY();
+            dualShipPosY = p2Icon->getPositionY() + 16.f;
+            dualUfoPosY = p2Icon->getPositionY() + 8.5f;
+            dualJetpackPosX = p2Icon->getPositionX() + 9.f;
+            dualJetpackPosY = p2Icon->getPositionY() + 6.f;
+            dualHasGlow = SDL->getSavedValue<bool>("glow");
+
+            IconPreview::updateDualPreviewCube(fields->m_dualPreview);
+
+            if (doGlowFix) {
+                if (lastDualMode == 1) {
+                    fields->m_dualGlowDisplay->setScale(1.6f);
+                    IconPreview::updateDualShipGlow(fields->m_dualGlowDisplay);
+
+                } else if (lastDualMode == 3) {
+                    fields->m_dualGlowDisplay->setScale(1.6f);
+                    IconPreview::updateDualBirdGlow(fields->m_dualGlowDisplay);
+
+                } else if (lastDualMode == 8) {
+                    fields->m_dualGlowDisplay->setScale(1.5f);
+                    IconPreview::updateDualJetpackGlow(fields->m_dualGlowDisplay);
+                }
+
+                fields->m_dualGlowDisplay->m_firstLayer->setOpacity(0);
+                fields->m_dualGlowDisplay->m_secondLayer->setVisible(false);
+                fields->m_dualGlowDisplay->m_birdDome->setVisible(false);
+                fields->m_dualGlowDisplay->m_detailSprite->setVisible(false);
+
+                if (lastDualMode == 1 || lastDualMode == 3 || lastDualMode == 8) {
+                    if (dualHasGlow) {
+                        fields->m_dualGlowDisplay->setVisible(true);
+                        fields->m_dualGlowDisplay->m_outlineSprite->setVisible(true);
+                        p2Icon->m_outlineSprite->setVisible(false);
+                    }
+                } else {
+                    fields->m_dualGlowDisplay->setVisible(false);
+                    p2Icon->m_outlineSprite->setVisible(dualHasGlow);
+                }                
+            }
+
+            if (lastDualMode == 1 || lastDualMode == 3 || lastDualMode == 8) {
+                fields->m_dualPreview->setVisible(true);
+            } else {
+                fields->m_dualPreview->setVisible(false);
+            }
+
+            if (lastDualMode == 1) { // SHIP
+                fields->m_dualPreview->runAction(CCEaseBackInOut::create(
+                    CCMoveTo::create(0.3f, {dualPosX, dualShipPosY})
+                ));
+                fields->m_dualPreview->runAction(CCEaseBackInOut::create(
+                    CCScaleTo::create(0.3f, 0.88f)
+                ));
+            } else if (lastDualMode == 3) { // UFO
+                fields->m_dualPreview->runAction(CCEaseBackInOut::create(
+                    CCMoveTo::create(0.3f, {dualPosX, dualUfoPosY})
+                ));
+                fields->m_dualPreview->runAction(CCEaseBackInOut::create(
+                    CCScaleTo::create(0.3f, 0.88f)
+                ));
+            } else if (lastDualMode == 8) { // JETPACK
+                fields->m_dualPreview->runAction(CCEaseBackInOut::create(
+                    CCMoveTo::create(0.3f, {dualJetpackPosX, dualJetpackPosY})
+                ));
+                fields->m_dualPreview->runAction(CCEaseBackInOut::create(
+                    CCScaleTo::create(0.3f, 0.9f)
+                ));
+            }
+
+        }
+
 
     }