+++ changelog.md
@@ -1,2 +1,8 @@
+# 1.0.1
+- Added **Jetpack Preview**: Cube now shows up on Jetpacks too!
+- Added *Separate Dual Icons* Compatibility.
+- Internal code improvements.
+- Added link to Source code (sorry).
+
 # 1.0.0
 - Initial release!